+++ src/export/executor.rs
@@ -1,5 +1,5 @@
 mod act_executor;
-// mod event_executor;
+mod event_executor;
 mod message_executor;
 mod model_executor;
 mod package_executor;
@@ -29,6 +29,7 @@ pub struct Executor {
     proc: process_executor::ProcessExecutor,
     task: task_executor::TaskExecutor,
     pack: package_executor::PackageExecutor,
+    evt: event_executor::EventExecutor,
 }
 
 impl ExecutorQuery {
@@ -86,6 +87,7 @@ impl Executor {
             proc: process_executor::ProcessExecutor::new(rt),
             task: task_executor::TaskExecutor::new(rt),
             pack: package_executor::PackageExecutor::new(rt),
+            evt: event_executor::EventExecutor::new(rt),
         }
     }
 
@@ -118,5 +120,10 @@ impl Executor {
     /// executor for related package functions
     pub fn pack(&self) -> &package_executor::PackageExecutor {
         &self.pack
+    }
+
+    /// executor for related event functions
+    pub fn evt(&self) -> &event_executor::EventExecutor {
+        &self.evt
     }
 }