+++ src/export/executor/event_executor.rs
@@ -1,7 +1,6 @@
 use crate::{
-    MessageInfo, Result, Vars,
-    scheduler::Runtime,
-    store::{PageData, StoreAdapter},
+    ActError, EventInfo, Result, Vars, scheduler::Runtime, store::PageData,
+    utils::consts,
 };
 use std::sync::Arc;
 use tracing::instrument;
@@ -19,28 +18,50 @@ impl EventExecutor {
             runtime: rt.clone(),
         }
     }
+
     #[instrument(skip(self))]
-    pub fn list(&self, q: &ExecutorQuery) -> Result<PageData<MessageInfo>> {
+    pub fn list(&self, q: &ExecutorQuery) -> Result<PageData<EventInfo>> {
         let query = q.into_query();
-        match self.runtime.cache().store().messages().query(&query) {
-            Ok(messages) => Ok(PageData {
-                count: messages.count,
-                page_size: messages.page_size,
-                page_count: messages.page_count,
-                page_num: messages.page_num,
-                rows: messages.rows.iter().map(|m| m.into()).collect(),
+        match self.runtime.cache().store().events().query(&query) {
+            Ok(events) => Ok(PageData {
+                count: events.count,
+                page_size: events.page_size,
+                page_count: events.page_count,
+                page_num: events.page_num,
+                rows: events.rows.iter().map(|m| m.into()).collect(),
             }),
             Err(err) => Err(err),
         }
     }
 
     #[instrument(skip(self))]
-    pub fn get(&self, id: &str) -> Result<MessageInfo> {
-        let message = &self.runtime.cache().store().messages().find(id)?;
-        Ok(message.into())
+    pub fn get(&self, id: &str) -> Result<EventInfo> {
+        let event = &self.runtime.cache().store().events().find(id)?;
+        Ok(event.into())
     }
 
-    pub fn do_event(&self, id: &str, options: &Vars) -> Result<()> {
-        Ok(())
+    pub fn start(&self, event_id: &str, params: &serde_json::Value) -> Result<Option<Vars>> {
+        let event = self.runtime.cache().store().events().find(event_id)?;
+
+        let register = self
+            .runtime
+            .package()
+            .get(&event.uses)
+            .ok_or(ActError::Runtime(format!(
+                "cannot find the registed package '{}'",
+                event.uses
+            )))?;
+
+        let options = Vars::new().with(consts::MODEL_ID, event.mid);
+
+        let mut params = params.clone();
+        if params.is_null() {
+            params = serde_json::from_str(&event.params).map_err(|err| {
+                ActError::Convert(format!("failed to deserialize params: {}", err))
+            })?;
+        }
+        let package = (register.create)(params)?;
+        let ret = package.start(&self.runtime, &options)?;
+        Ok(ret)
     }
 }