+++ src/package/tests/mod.rs
@@ -1,6 +1,7 @@
 mod action;
 mod block;
 mod code;
+mod event;
 mod irq;
 mod msg;
 mod parallel;