+++ src/store/db/mem/impl/event.rs
@@ -15,7 +15,9 @@ impl DbDocument for Event {
         map.insert("id".to_string(), json!(self.id.clone()));
         map.insert("name".to_string(), json!(self.name.clone()));
         map.insert("mid".to_string(), json!(self.mid.clone()));
-
+        map.insert("ver".to_string(), json!(self.ver.clone()));
+        map.insert("uses".to_string(), json!(self.uses.clone()));
+        map.insert("params".to_string(), json!(self.params.clone()));
         map.insert("create_time".to_string(), json!(self.create_time));
         map.insert("timestamp".to_string(), json!(self.timestamp));
         Ok(map)